Refactor: remove easter egg function to separate file
